// types/StyleType.tsx
export enum ColorHexProps {
    black = "black",
    white = "white",
    blue = "#3D5AFE",
    lightBlue = "#2962FF",
    persianBlue ="#0039CB",
    green = "#38b000",
    deepGreen = "#007200",
    pink = "#ff85a1",
    deepPink = "#ff5c8a",
    red = "#e5383b",
    deepRed = "#ba181b",
    pennRed = "#9A0007",
    persianRed = "#D32F2F",
    orange = "#fb8500",
    deepOrange = "#e85d04",
    yellow = "#ffea00",
    deepYellow = "#ffb700",
    purple = "#a06cd5",
    deepPurple = "#6247aa",
    gray = "#E0E0E0",
    deepGray = "#3F3F3F",
    lightGray = "#AEAEAE",
    cadetGray = "#9E9E9E",
    jetGray = "#333",
    paynesGray = "#455A64",
    gunmetal = "#1C313A",
    mintCream = "#EFF7F6",
}

export enum ColorProps {
    black = "0, 0, 0",
    white = "255, 255, 255",
    blue = "61, 90, 254",
    lightBlue = "41, 98, 255",
    persianBlue ="0, 49, 202",
    green = "56, 176, 0",
    deepGreen = "0, 114, 0",
    pink = "255, 133, 161",
    deepPink = "255, 92, 138",
    red = "229, 56, 59",
    deepRed = "186, 24, 27",
    orange = "232, 93, 4",
    deepOrange = "232, 93, 4",
    yellow = "255, 234, 0",
    deepYellow = "255, 183, 0",
    purple = "160, 108, 213",
    deepPurple = "98, 71, 170",
    gray = "224, 224, 224",
    deepGray = "63, 63, 63",
    lightGray = "174, 174, 174",
    cadetGray = "158, 158, 158",
    jetGray = "51, 51, 51",
    paynesGray = "69, 90, 100",
    gunmetal = "28, 49, 58",
    pennRed = "154, 0, 7",
    persianRed = "211, 47, 47",
    mintCream = "239, 247, 246",
}